requirements.txt
altgraph==0.17.2
bcrypt==3.2.0
cffi==1.15.0
cryptography==36.0.1
dnspython==2.2.0
macholib==1.15.2
paramiko==2.9.2
pycparser==2.21
pyinstaller==4.8
pyinstaller-hooks-contrib==2021.5
PyJWT==2.3.0
pymongo==3.11.3
PyNaCl==1.5.0
pysftp==0.2.9
python-dotenv==0.19.2
six==1.16.0
pytz==2021.3
requests==2.27.1
numpy==1.22.3
pandas==1.4.1
